Move Match entity into MatchAggregate
