+++ src/TournamentManagement/TournamentManagement.Domain.UnitTests/MatchAggregate/MatchTests.cs
@@ -1,10 +1,11 @@
 ﻿using FluentAssertions;
 using System;
 using System.Collections.Generic;
+using TournamentManagement.Domain.MatchAggregate;
 using TournamentManagement.Domain.TournamentAggregate;
 using Xunit;
 
-namespace TournamentManagement.Domain.UnitTests
+namespace TournamentManagement.Domain.UnitTests.MatchAggregate
 {
 	public class MatchTests
 	{
