+++ src/TournamentManagement/TournamentManagement.Domain.UnitTests/MatchAggregate/SetWinnerCalculatorTests.cs
@@ -1,10 +1,10 @@
 ﻿using FluentAssertions;
-using System;
+using TournamentManagement.Domain.MatchAggregate;
 using TournamentManagement.Domain.TournamentAggregate;
 using Xunit;
 
 
-namespace TournamentManagement.Domain.UnitTests
+namespace TournamentManagement.Domain.UnitTests.MatchAggregate
 {
 	public class SetWinnerCalculatorTests
 	{
