+++ src/TournamentManagement/TournamentManagement.Domain/MatchAggregate/MatchOutcome.cs
@@ -1,4 +1,4 @@
-﻿namespace TournamentManagement.Domain
+﻿namespace TournamentManagement.Domain.MatchAggregate
 {
 	public enum MatchOutcome
 	{
