+++ src/TournamentManagement/TournamentManagement.Domain/MatchAggregate/MatchState.cs
@@ -1,4 +1,4 @@
-﻿namespace TournamentManagement.Domain
+﻿namespace TournamentManagement.Domain.MatchAggregate
 {
 	public enum MatchState
 	{
